+++ lib/dropdown_search.dart
@@ -153,6 +153,10 @@ class DropdownSearch<T> extends StatefulWidget {
   ///custom clear button widget builder
   final IconButtonBuilder? clearButtonBuilder;
 
+  ///custom splash radius for the clear button
+  ///If null, default splash radius of [icon_button] is used.
+  final double? clearButtonSplashRadius;
+
   ///custom dropdown icon button widget
   final Widget? dropDownButton;
 
@@ -162,6 +166,10 @@ class DropdownSearch<T> extends StatefulWidget {
 
   ///custom dropdown button widget builder
   final IconButtonBuilder? dropdownButtonBuilder;
+
+  ///custom splash radius for the dropdown button
+  ///If null, default splash radius of [icon_button] is used.
+  final double? dropdownButtonSplashRadius;
 
   ///whether to manage the clear and dropdown icons via InputDecoration suffixIcon
   final bool showAsSuffixIcons;
@@ -250,8 +258,10 @@ class DropdownSearch<T> extends StatefulWidget {
     this.dialogMaxWidth,
     this.clearButton,
     this.clearButtonBuilder,
+    this.clearButtonSplashRadius,
     this.dropDownButton,
     this.dropdownButtonBuilder,
+    this.dropdownButtonSplashRadius,
     this.showAsSuffixIcons = false,
     this.dropdownBuilderSupportsNullItem = false,
     this.popupShape,
@@ -412,6 +422,7 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
               : IconButton(
                   icon: widget.clearButton ?? const Icon(Icons.clear, size: 24),
                   onPressed: clearButtonPressed,
+                  splashRadius: widget.clearButtonSplashRadius ?? null,
                 ),
         widget.dropdownButtonBuilder != null
             ? InkWell(
@@ -422,6 +433,7 @@ class DropdownSearchState<T> extends State<DropdownSearch<T>> {
                 icon: widget.dropDownButton ??
                     const Icon(Icons.arrow_drop_down, size: 24),
                 onPressed: dropdownButtonPressed,
+                splashRadius: widget.dropdownButtonSplashRadius ?? null,
               ),
       ],
     );